# docs/bix-openapi.yaml
openapi: 3.0.3
info:
  title: BIX — Ghaf Business Intelligence API
  version: 1.0.0
  description: |
    Complete reverse-engineered API surface of the BIX healthcare revenue-cycle
    management solution (ASP.NET Core 10 MVC, SQLite, ClosedXML, Hangfire-lite,
    dual AI providers). Generated from source analysis of all 9 controllers +
    mapped endpoints.

    ## Conventions
    - **Routing**: conventional MVC `/{Controller}/{Action}` unless an explicit
      route is shown. `ReportScheduler` uses attribute routing
      `[Route("[controller]/[action]")]` (same shape).
    - **Auth**: ASP.NET Identity cookie session. Unauthenticated requests to
      protected endpoints return **302 → /login** (not 401).
    - **CSRF**: every state-changing POST validates an anti-forgery token —
      include form field `__RequestVerificationToken` (from any rendered page).
      `Support/Chat` is the JSON-body exception (still cookie-authed).
    - **HTML endpoints** render Razor views; they're documented tersely with
      `text/html` responses. JSON/stream/file endpoints are fully typed.
    - **SSE endpoints** return `text/event-stream` of `data: <json>` frames.

    ## Roles
    | Policy | Roles |
    |---|---|
    | AdminAccess | Admin |
    | RcmAccess | Admin, FacilityAdmin, Analyst, Billing, Finance, Auditor |
    | ReportAccess | RcmAccess + Reporter |
    | ResubmissionAccess | Admin, FacilityAdmin, Analyst |
servers:
  - url: https://bix.ghafservices.com
...
    description: Local
...
security:
  - cookieAuth: []
paths:
...
  /login:
    get:
      tags: [Auth]
      security: []
      summary: Login page (aliases /Home/Index, /Home/Login)
      responses: { "200": { description: HTML login page } }
    post:
      tags: [Auth]
      security: []
      summary: Sign in (aliases /Home/Index, /Home/Login)
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [Email, Password, __RequestVerificationToken]
              properties:
                Email: { type: string, format: email }
                Password: { type: string, format: password }
                RememberMe: { type: boolean }
                __RequestVerificationToken: { type: string }
      responses:
        "302": { description: Redirect to dashboard on success; back to login on failure }
...
  /api/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Cached dashboard summary JSON (keepalive/pre-warm target)
      responses:
        "200": { description: Summary metrics JSON, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "302": { $ref: "#/components/responses/LoginRedirect" }

  # ───────────────────────────── AI Assistant ─────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application
      description: ASP.NET Identity session cookie (8h sliding). Login via POST /login.
  responses:
    LoginRedirect:
      description: Not authenticated / insufficient role — 302 redirect to /login
...